Storing blood typing test results in the database Yet to implement logic to avoid duplicate entry of results for the same collection Also avoid using TINYINT

+++ src/model/bloodbagtype/BloodBagType.java
@@ -14,7 +14,7 @@ public class BloodBagType {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="TINYINT")
+  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="SMALLINT")
   private Integer id;
 
   @Column(length=50)
