+++ src/model/bloodtyping/rules/BloodTypingRule.java
@@ -18,7 +18,7 @@ public class BloodTypingRule {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="TINYINT")
+  @Column(nullable = false, insertable=false, updatable=false, columnDefinition="MEDIUMINT")
   private Integer id;
 
   @Column(length=200)
