+++ src/model/donationtype/DonationType.java
@@ -14,7 +14,7 @@ public class DonationType {
 
   @Id
   @GeneratedValue(strategy = GenerationType.AUTO)
-  @Column(nullable=false, insertable=false, updatable=false, columnDefinition="TINYINT")
+  @Column(nullable=false, insertable=false, updatable=false, columnDefinition="SMALLINT")
   private Integer id;
 
   @Column(length=50)
